+++ src/main.cpp
@@ -2,6 +2,8 @@
 
 using namespace geode::prelude;
 
+static bool special = false;
+
 #include <Geode/modify/EditLevelLayer.hpp>
 class $modify(EditLevelLayer) {
 	bool init(GJGameLevel* p) {
@@ -31,17 +33,28 @@ class $modify(EditLevelLayer) {
 		menu->addChild(btn);
 
 		// transparent
-		if (Mod::get()->getSettingValue<bool>("transparent")){
+		if (Mod::get()->getSettingValue<bool>("transparent")) {
 			this->getChildByID("level-name-background")->setVisible(false);
 			this->getChildByID("description-background")->setVisible(false);
 		}
-
 		return true;
 	}
 };
 
+#include <Geode/modify/GJGarageLayer.hpp>
+class $modify(MyGarageLayer, GJGarageLayer) {
+public:
+	void onBack(CCObject* sender) {
+		if (special)
+			CCDirector::get()->popSceneWithTransition(0.5f, PopTransition::kPopTransitionFade);
+		else
+			GJGarageLayer::onBack(sender);
+		special = false;
+	}
+};
+
 #include <Geode/modify/LevelSelectLayer.hpp>
-class $modify(LevelSelectLayer) {
+class $modify(OfficialBrowser, LevelSelectLayer) {
 	bool init(int page) {
 		if (!LevelSelectLayer::init(page))
 			return false;
@@ -64,7 +77,7 @@ class $modify(LevelSelectLayer) {
 		auto btn = CCMenuItemSpriteExtra::create(
 			spr,
 			this,
-			menu_selector(MenuLayer::onGarage));
+			menu_selector(OfficialBrowser::onGarage));
 		btn->m_animationType = MenuAnimationType::Move;
 		btn->m_startPosition = CCPoint(18.875, 35.25);
 		btn->m_offset = CCPoint(0.f, -8.f);
@@ -72,28 +85,11 @@ class $modify(LevelSelectLayer) {
 		
 		return true;
 	}
-};
 
-/*
-class $modify(LevelInfoLayer) {
-	bool init(GJGameLevel* p0, bool p1) {
-		if (!LevelInfoLayer::init(p0, p1)) {
-			return false;
-			}
-
-		auto menu = static_cast<CCMenuItemSpriteExtra* >(this->getChildByID("garage-menu")->getChildByID("garage-button"));
-		auto num1 = menu->m_destPosition.x;
-		auto num2 = menu->m_destPosition.y;
-		auto num3 = menu->m_startPosition.x;
-		auto num4 = menu->m_startPosition.y;
-		auto num5 = menu->m_offset.x;
-		auto num6 = menu->m_offset.y;
-	fstream myFile;
-	myFile.open("D:/Alvorika my waifu.txt", ios::out);
-	if (myFile.is_open())
-	{
-		myFile << num1 << '\n' << num2 << '\n' << num3 << '\n' << num4 << '\n' << num5 << '\n' << num6 << '\n' << '\n';
-		myFile.close();
-	}return true;
+	void onGarage(CCObject* sender) {
+		special = true;
+		auto scene = MyGarageLayer::scene();
+		auto effect = CCTransitionMoveInT::create(0.4, scene);
+		CCDirector::sharedDirector()->pushScene(effect);
 	}
-};*/
+};