+++ changelog.md
@@ -1,3 +1,7 @@
+# 3.3.2
+Make the transition of LevelSelectLayer's garage rope more human-like;  
+and the garage menu from LevelSelectLayer will return to LevelSelectLayer instead of MenuLayer (main menu) from now on  
+  
 # 3.3.1
 Fix the LevelSelectLayer button usually not working